fix(ci): add test client_id into fhir.smart.service.ts file

+++ src/app/common/fhir/smart/services/fhir.smart.service.ts
@@ -72,6 +72,10 @@ export class FhirSmartService {
       console.error('context is not supported', context);
       return;
     }
+    if (!environment.client_id[context]) {
+      console.error('client_id is not supported for context', context);
+      return;
+    }
     sessionStorage.clear();
     sessionStorage.setItem('context', context);
     this.setBaseUrl(iss);
